Make some requirements less requirementy
The PackageRepo and ShellInstaller requirements would always run,
regardless of whether or not they needed to.
For the PackageRepo, check to see if the apt repo is already in the
package lists; if so then don't do the expensive add and apt-get update.

For the ShellInstaller, since the installer doesn't really know what
it's installing, we allow it to provide a shell test.

There's also a BREAKING CHANGE here: the ShellInstaller now by default
uses run() rather than sudo() to make it consistent with the other
requirements. Use use_sudo=True to use sudo().